fixed message pop bug
